requirements.txt
django>=1.7.1,<1.8
django-bootstrap3
djangorestframework
django-filter
markdown
psycopg2
rest-pandas
